+++ apps/web/src/components/Nfts/NftFeed.tsx
@@ -1,4 +1,4 @@
-import SingleNft from '@components/NFT/SingleNft';
+import SingleNft from '@components/Nfts/SingleNft';
 import NFTSShimmer from '@components/Shared/Shimmer/NFTSShimmer';
 import { EmptyState } from '@components/UI/EmptyState';
 import { ErrorMessage } from '@components/UI/ErrorMessage';
