+++ apps/web/src/components/Profile/NftGallery/Picker.tsx
@@ -1,4 +1,4 @@
-import SingleNft from '@components/NFT/SingleNft';
+import SingleNft from '@components/Nfts/SingleNft';
 import NftPickerShimmer from '@components/Shared/Shimmer/NftPickerShimmer';
 import { EmptyState } from '@components/UI/EmptyState';
 import { ErrorMessage } from '@components/UI/ErrorMessage';
